Add pages for create Categories and Products

+++ backend/controllers/Product/productsController.js
@@ -1,49 +1,84 @@
 const db = require('../../models/index.js');
-
+const path = require('path');
 const Product = db.products;
+const Category = db.category;
 const User = db.user;
 
+function makeRelativeFilesRoute(files){
+    // This make list route relative files static for save in model Products
+    const listSeparateRoutes = [];
+    for(let i=0; i <files.length;i++ ){
+        const splitRoute = files[i].destination.split(path.sep).splice(-2)
+        splitRoute.push(files[i].filename)
+        const relativePath  = `${splitRoute[0]}/${splitRoute[1]}/${splitRoute[2]}`
+        listSeparateRoutes.push(relativePath);
+    }
+    return listSeparateRoutes;
+}
 const addProduct = async (req, res)=>{
 
-    const user_id = req.user.id;
-    
-    let values = {
-        name : req.body.name,
-        description : req.body.desc,
-        specification: req.body.spec,
-        price: req.body.price,
-        old_price : req.body.oldPrice,
-        available: req.body.available,
-        user_id
+    try{
+        const userEmail = req.user.email
+        // Find id user request.
+        const user = await User.findOne({where:{email:userEmail}})
+        if(user.role !== 'admin' && user.role !== 'moderator'){
+            return res.status(401).json({message:"You do not have authorization to create this product."})
+        }
+        // validation field images.
+        const files = req.files;
+        if(!files.length){
+            return res.status(400).json({message:"Requires at least one image to create this product."})
+        }
+        const values = req.body;
+        const categories = values.categories.split(',').map(id=>parseInt(id));
+        const relativeRouteFilestoStr = makeRelativeFilesRoute(files).toString();
+        const images = relativeRouteFilestoStr;
+
+        const newValues = {
+            name : values.name,
+            description: values.description,
+            specification: values.specification,
+            price : values.price,
+            old_price: values.oldPrice,
+            available: values.available,
+            user_id: user.id,
+            images,
+        }
+
+        const product = await Product.create(newValues).then(product=>{
+            product.addCategories(categories);
+            return product;
+        })
+
+        res.status(200).json({message:'Product add successfully'});
+    }
+    catch(err){
+        console.log('Something Wrong in create product');
+        res.status(500).json({"message":"Server Error"});
     }
 
-    try{
-        const product = await Product.create(values);
-        
-        res.status(200).json(product);
-    }
-    catch{
-        console.log('Something Wrong in create product')
-        res.status(500).json({"message":"Server Error"})
-    }
 }
 
 const getAllProducts = async (req,res)=>{
     try{
         let products = await Product.findAll({
             attributes:{
-                exclude:['updatedAt','createdAt','categoriumId','user_id']
+                exclude:['updatedAt','user_id']
             },
-        //     include: [
-        //         {
-        //             association:Product.User,
-        //             attributes:{
-        //                 exclude:['password','is_active','role','updatedAt','createdAt','UUID']
-        //             }
-
-        //         }
-        // ]
+            order:[
+                ['createdAt','DESC']
+            ],
+            include: [
+                {
+                    association:Product.User,
+                    as:'user',
+                    attributes:{
+                        exclude:['password','is_active','role','updatedAt','createdAt','UUID']
+                    }
+                }]
         });
+
+
         res.status(200).json(products);
     }catch(err){
         console.log('Something Wrong in get all products');
@@ -98,9 +133,16 @@ const removeProduct = async (req,res)=>{
     try{
 
         let id = req.params.id
-        const product = Product.destroy({where: {id: id}})
+        const userEmail = req.user.email
+        // Find id user request.
+        const user = await User.findOne({where:{email:userEmail}})
+        if(user.role !== 'admin' && user.role !== 'moderator'){
+            return res.status(401).json({message:"You do not have authorization to delete this product."})
+        }
+
+        const product = await Product.destroy({where: {id: id}})
         if(!product){
-            res.status(401).json({message:"The product does not exist"})
+            return res.status(401).json({message:"The product does not exist"})
         }
         res.status(200).json({"message":"Product deleted successfully."});
     }catch{
